Move the model related classes to the Model package
