+++ app/src/main/java/com/arctouch/codechallenge/model/data/Cache.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.data
+package com.arctouch.codechallenge.model.data
 
 import com.arctouch.codechallenge.model.Genre
